+++ bakedanuki/bakedanuki-util/python/bd_util/ui/__init__.py
@@ -55,6 +55,7 @@ from .binding import (
 )
 from .settings_path import SettingsPath
 from .json_clipboard import JsonClipboard
+from .float_step_profile import FloatStepProfile, FloatStepSetting
 from .radio_button_sweep import RadioButtonSweep
 from .check_box_sweep import CheckBoxSweep
 from .ui_state import UiStateManager
@@ -67,6 +68,8 @@ from .window_state import (
 
 __all__ = [
     "JsonClipboard",
+    "FloatStepProfile",
+    "FloatStepSetting",
     "CheckBoxSweep",
     "RadioButtonSweep",
     "EnumBinding",